RT-15 Break out of remaining ELK roles
As part of the effort to move ELK upstream this commit removes ELK
roles in the RPC-O tree in favor of the new rcbops repos.  This commit
also upgrades ELK versions to the 5.x release.

1. Move to fork of ansible-elasticsearch

The ansible-elasticsearch from elastic.co uses an attribute called
check_mode which is not supported in ansible < 2.2.x, since we're
using 2.1.5 we need to remove that attribute.  Once RPC-O is running
ansible > 2.2.x we can revert this change.

2. Add log_hosts configuration for kernel parameters

The ELK stack needs some specific kernel parameters to be at certain
minimum values in order to start. We create a log_hosts.yml and override
the kernel parameters to get ELK started.

3. Add new ELK role names to artifact build process

Make use of the new roles when building the artifacts.