RE-191 Add testing flag to migrate-yaml.py
For gating purposes, we need a flag in scripts/migrate.yml.py that does
the opposite of --for-testing-take-new-vars-only.  When we do a major
upgrade, for example, we do not re-run deploy.sh and therefore we do
not want to wipe out our previously defined overrides.  This commit
adds a new flag called --for-testing-keep-old-overrides-only which
keeps the old overrides rather than the new conflicting defaults.

Note that this was never an issue in gating as we wrote gating-specific
overrides to a gating overrides file which was never touched by the
upgrade process.  However, as we move away from making gating-related
changes we need to handle this properly in rpc-openstack itself.

This change is necessary to unblock a ceph gating major upgrade gating
failure as the upgrade is wiping out some flags which we used to
override in rpc-gating itself.